refactor: split FindingCategoryHotspot into focused modules
Extract HotspotGrid (with HotspotLegend) and all pure color/data
utilities into separate files to reduce the main widget file from
434 lines to ~207 lines.

- FindingCategoryHotspotUtils.js: color constants and pure functions
  (getTop10, computeLegendRanges, getColorForValue, buildLegendLabels,
  getTextColorForLevel)
- HotspotGrid.js: HotspotLegend + HotspotGrid presentational components
- FindingCategoryHotspot.js: main widget only; re-exports utility
  functions so existing test imports remain unchanged

All 33 existing tests pass.
